refactoring

+++ parser_library/src/workspaces/file_impl.cpp
@@ -235,7 +235,7 @@ size_t file_impl::index_from_position(const std::string& text, const std::vector
     size_t end = (size_t)loc.column;
     if (loc.line >= line_indices.size())
         return text.size();
-    size_t i = line_indices[(size_t)loc.line];
+    size_t i = line_indices[loc.line];
     size_t utf16_counter = 0;
 
     while (utf16_counter < end && i < text.size())
